requirements.txt
appdirs==1.4.4
asgiref==3.2.3
blis==0.4.1
cached-property==1.5.1
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cymem==2.0.3
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==2.2.10
django-appconf==1.0.3
django-braces==1.14.0
django-common-helpers==0.9.2
django-cors-headers==3.2.0
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
easy-thumbnails==2.6
fastapi==0.45.0
gunicorn==20.0.4
h11==0.9.0
httpie==2.2.0
idna==2.8
importlib-metadata==1.7.0
joblib==0.16.0
Markdown==3.2.2
more-itertools==8.4.0
murmurhash==1.0.2
ordered-set==3.1.1
ordered-set-stubs==0.1.3
pdfkit==0.6.1
Pillow==6.2.1
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.2
psycopg2==2.8.4
psycopg2-binary==2.8.5
pycparser==2.20
pydantic==1.6.1
Pygments==2.6.1
PyJWT==1.7.1
python-decouple==3.3
python-http-client==3.2.7
python3-openid==3.2.0
pytz==2019.3
redis==3.5.3
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.15.100
six==1.13.0
sqlparse==0.3.0
srsly==1.0.2
starkbank-ecdsa==1.0.0
starlette==0.12.9
thinc==7.4.1
threadpoolctl==2.1.0
tomlkit==0.5.11
tqdm==4.48.0
urllib3==1.25.7
uvicorn==0.11.6
wasabi==0.7.1
websockets==8.1
whitenoise==4.1.4
wkhtmltopdf==0.2
xlrd==1.2.0